+++ src/service/globals/unpublish.rs
@@ -87,7 +87,7 @@ pub fn unpublish_global_document(ctx: &ServiceContext) -> Result<Document> {
     query::hydrate_document(&tx, &gtable, &def.fields, &mut doc, None, None)?;
 
     let mut read_denied = wh.field_read_denied(&def.fields, ctx.user);
-    read_denied.extend(helpers::collect_hidden_field_names(&def.fields, ""));
+    read_denied.extend(helpers::collect_api_hidden_field_names(&def.fields, ""));
 
     doc.strip_fields(&read_denied);